Modularize credentials.rs into submodules for maintainability
Split 504-line credentials.rs into focused submodules:
- mod.rs: Credential type and re-exports (28 lines)
- storage.rs: File storage operations and test mocks (167 lines)
- migration.rs: Keyring migration logic (128 lines)
- tests.rs: All unit tests (191 lines)
